Minor service refactor
