+++ sources/packages/backend/apps/api/src/route-controllers/cas-invoice-batch/cas-invoice-batch.aest.controller.ts
@@ -11,7 +11,10 @@ import { AuthorizedParties, Role, UserGroups } from "../../auth";
 import { AllowAuthorizedParty, Groups, Roles } from "../../auth/decorators";
 import BaseController from "../BaseController";
 import { ClientTypeBaseRoute } from "../../types";
-import { CASInvoiceBatchService } from "../../services";
+import {
+  CASInvoiceBatchReportService,
+  CASInvoiceBatchService,
+} from "../../services";
 import { CASInvoiceBatchAPIOutDTO } from "./models/cas-invoice-batch.dto";
 import { getUserFullName } from "../../utilities";
 import {
@@ -28,7 +31,10 @@ import { streamFile } from "../utils";
 @Controller("cas-invoice-batch")
 @ApiTags(`${ClientTypeBaseRoute.AEST}-cas-invoice-batch`)
 export class CASInvoiceBatchAESTController extends BaseController {
-  constructor(private readonly casInvoiceBatchService: CASInvoiceBatchService) {
+  constructor(
+    private readonly casInvoiceBatchService: CASInvoiceBatchService,
+    private readonly casInvoiceBatchReportService: CASInvoiceBatchReportService,
+  ) {
     super();
   }
 
@@ -70,7 +76,7 @@ export class CASInvoiceBatchAESTController extends BaseController {
     @Res() response: Response,
   ): Promise<void> {
     const invoiceReport =
-      await this.casInvoiceBatchService.getCASInvoiceBatchesReport(
+      await this.casInvoiceBatchReportService.getCASInvoiceBatchesReport(
         casInvoiceBatchId,
       );
     const batchDate = getFileNameAsCurrentTimestamp(invoiceReport.batchDate);
