+++ sources/packages/backend/apps/api/src/services/index.ts
@@ -52,3 +52,4 @@ export * from "./audit/audit.service";
 export * from "./audit/audit-event.enum";
 export * from "./application-restriction-bypass/application-restriction-bypass.models";
 export * from "./cas-invoice-batch/cas-invoice-batch.service";
+export * from "./cas-invoice-batch/cas-invoice-batch-report.service";
